+++ changelog.md
@@ -17,6 +17,10 @@
 - [Chrome - Ajout du lien de l'extension](https://github.com/RudyLaban/JPlugCheck/commit/b2da38d61deceb9c79580cbdc6f7b9ce2f17ea73)
 - [Reformulation et ajout à la doc](https://github.com/RudyLaban/JPlugCheck/commit/4e745b2c961f56fbf491764aa33620000c9ac819)
 
+### Fixes
+
+- [Popup - Bloc "done" affiché alors que traitement en cours]()
+
 ---
 
 ## 1.1 - Publiée (16.06.25)